requirements.txt
astroid==2.4.2
autopep8==1.5.4
backcall==0.2.0
decorator==4.4.2
distro==1.5.0
ipykernel==5.3.4
ipython==7.19.0
ipython-genutils==0.2.0
isort==5.6.4
jedi==0.17.2
jupyter-client==6.1.7
jupyter-core==4.7.0
lazy-object-proxy==1.4.3
lxml==4.6.2
mccabe==0.6.1
numpy==1.19.4
pandas==1.1.4
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
pikepdf==2.2.0
Pillow==8.0.1
prompt-toolkit==3.0.8
ptyprocess==0.6.0
pycodestyle==2.6.0
Pygments==2.7.2
pylint==2.6.0
PyPDF2==1.26.0
python-dateutil==2.8.1
pytz==2020.4
pyzmq==20.0.0
six==1.15.0
tabula-py==2.2.0
toml==0.10.2
tornado==6.1
traitlets==5.0.5
wcwidth==0.2.5
wrapt==1.12.1
